refactor: FSD- entities 배치하기
- api: api와 관련된 타입
- model: 엔티티의 타입
