+++ src/entities/comments/api/CommentsResponse.ts
@@ -1,0 +1,3 @@
+import { ApiResponseList } from '../../../shared/api';
+
+export type CommentsResponse = ApiResponseList<Comment, 'comments'>;
