+++ src/entities/posts/api/PostsResponse.ts
@@ -1,0 +1,4 @@
+import { ApiResponseList } from '../../../shared/api';
+import { Post } from '../model/Post';
+
+export type PostsResponse = ApiResponseList<Post, 'posts'>;
