+++ src/entities/posts/model/NewPost.ts
@@ -1,0 +1,3 @@
+import { Post } from './Post';
+
+export type NewPost = Pick<Post, 'title' | 'body' | 'userId'>;
