+++ src/entities/posts/model/index.ts
@@ -1,0 +1,3 @@
+export type { Post } from './Post';
+export type { NewPost } from './NewPost';
+export type { PostsUrlParams } from './PostUrlParams';
