+++ src/entities/tags/model/Tag.ts
@@ -1,0 +1,4 @@
+export interface Tag {
+  slug: string;
+  url: string;
+}
