+++ src/entities/users/model/User.ts
@@ -1,0 +1,5 @@
+export interface User {
+  id: number;
+  username: string;
+  image: string;
+}
