--- src/types/index.ts
@@ -1,4 +1,0 @@
-export type { User, UserResponse, UsersResponse } from './User';
-export type { Post, PostsResponse } from './Post';
-export type { Comment, Comments } from './Comment';
-export type { Tag, Tags } from './Tag';
